+++ src/main/java/com/braintreegateway/CreditCardVerification.java
@@ -23,6 +23,8 @@ public class CreditCardVerification {
     private String processorResponseCode;
     private String processorResponseText;
     private ProcessorResponseType processorResponseType;
+    private String networkResponseCode;
+    private String networkResponseText;
     private String merchantAccountId;
     private Status status;
     private String id;
@@ -42,6 +44,8 @@ public class CreditCardVerification {
         this.processorResponseCode = node.findString("processor-response-code");
         this.processorResponseText = node.findString("processor-response-text");
         this.processorResponseType = EnumUtils.findByName(ProcessorResponseType.class, node.findString("processor-response-type"), ProcessorResponseType.UNRECOGNIZED);
+        this.networkResponseCode = node.findString("network-response-code");
+        this.networkResponseText = node.findString("network-response-text");
         this.merchantAccountId = node.findString("merchant-account-id");
         this.status = EnumUtils.findByName(Status.class, node.findString("status"), Status.UNRECOGNIZED);
         this.id = node.findString("id");
@@ -123,6 +127,14 @@ public class CreditCardVerification {
 
     public ProcessorResponseType getProcessorResponseType() {
         return processorResponseType;
+    }
+
+    public String getNetworkResponseCode() {
+        return networkResponseCode;
+    }
+
+    public String getNetworkResponseText() {
+        return networkResponseText;
     }
 
     public String getMerchantAccountId() {