+++ src/test/java/com/braintreegateway/integrationtest/CreditCardVerificationIT.java
@@ -54,6 +54,42 @@ public class CreditCardVerificationIT extends IntegrationTest {
     }
 
     @Test
+    public void createVerificationNetworkResponseCodeText() {
+        CreditCardVerificationRequest request = new CreditCardVerificationRequest().
+            creditCard().
+                number(CreditCardNumber.VISA.number).
+                expirationDate("05/2009").
+                cvv("123").
+                billingAddress().
+                    company("Braintree").
+                    countryCodeAlpha2("US").
+                    countryCodeAlpha3("USA").
+                    countryCodeNumeric("840").
+                    countryName("United States of America").
+                    extendedAddress("Unit B").
+                    firstName("John").
+                    lastName("Smith").
+                    locality("San Francisco").
+                    postalCode("60606").
+                    region("CA").
+                    streetAddress("123 Townsend St").
+                    done().
+                done().
+            options().
+                amount("5.00").
+                done();
+
+        Result<CreditCardVerification> result = gateway.creditCardVerification().create(request);
+        assertTrue(result.isSuccess());
+        CreditCardVerification verification = result.getTarget();
+        assertEquals("1000", verification.getProcessorResponseCode());
+        assertEquals("Approved", verification.getProcessorResponseText());
+        assertEquals(ProcessorResponseType.APPROVED, verification.getProcessorResponseType());
+        assertEquals("XX", verification.getNetworkResponseCode());
+        assertEquals("sample network response text", verification.getNetworkResponseText());
+    }
+
+    @Test
     public void createVerificationDeclined() {
         CreditCardVerificationRequest request = new CreditCardVerificationRequest().
             creditCard().